+++ CMakeLists.txt
@@ -3,4 +3,8 @@ project(C_Calculator C)
 
 set(CMAKE_C_STANDARD 11)
 
-add_executable(C_Calculator main.c)
+add_executable(C_Calculator main.c
+        arithmetic.c
+        menuHandler.c
+        menuHandler.h
+        arithmetic.h)
