--- cmake-build-debug/.cmake/api/v1/reply/codemodel-v2-bb4466bcb330aba4ef41.json
@@ -1,60 +1,0 @@
-{
-	"configurations" : 
-	[
-		{
-			"directories" : 
-			[
-				{
-					"build" : ".",
-					"jsonFile" : "directory-.-Debug-d0094a50bb2071803777.json",
-					"minimumCMakeVersion" : 
-					{
-						"string" : "3.28"
-					},
-					"projectIndex" : 0,
-					"source" : ".",
-					"targetIndexes" : 
-					[
-						0
-					]
-				}
-			],
-			"name" : "Debug",
-			"projects" : 
-			[
-				{
-					"directoryIndexes" : 
-					[
-						0
-					],
-					"name" : "C_Calculator",
-					"targetIndexes" : 
-					[
-						0
-					]
-				}
-			],
-			"targets" : 
-			[
-				{
-					"directoryIndex" : 0,
-					"id" : "C_Calculator::@6890427a1f51a3e7e1df",
-					"jsonFile" : "target-C_Calculator-Debug-a8f8eba4879c40d818ea.json",
-					"name" : "C_Calculator",
-					"projectIndex" : 0
-				}
-			]
-		}
-	],
-	"kind" : "codemodel",
-	"paths" : 
-	{
-		"build" : "C:/Users/domir/CLionProjects/C_Calculator/cmake-build-debug",
-		"source" : "C:/Users/domir/CLionProjects/C_Calculator"
-	},
-	"version" : 
-	{
-		"major" : 2,
-		"minor" : 6
-	}
-}
