+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: May 02 20:28 Eastern Daylight Time
+Start testing: May 16 20:52 Eastern Daylight Time
 ----------------------------------------------------------
-End testing: May 02 20:28 Eastern Daylight Time
+End testing: May 16 20:52 Eastern Daylight Time
