+++ menuHandler.h
@@ -1,0 +1,10 @@
+//
+// Created by domir on 5/16/2024.
+//
+
+#ifndef C_CALCULATOR_MENUHANDLER_H
+#define C_CALCULATOR_MENUHANDLER_H
+
+void menu(void);
+
+#endif //C_CALCULATOR_MENUHANDLER_H
